poc/t1_risk_verification/requirements.txt
google-cloud-secret-manager==2.18.1
google-auth==2.23.4
google-cloud-core==2.3.3